requirements.txt
requests
python-cfonts
termcolor
